+++ src/dragen_align_pa/jobs/manage_dragen_mlr.py
@@ -12,6 +12,7 @@ from cpg_utils.hail_batch import get_batch
 from hailtop.batch.job import PythonJob
 from loguru import logger
 
+from dragen_align_pa import ica_utils
 from dragen_align_pa.constants import BUCKET_NAME, ICA_CLI_SETUP
 from dragen_align_pa.jobs.ica_pipeline_manager import manage_ica_pipeline_loop
 
@@ -23,77 +24,6 @@ def _initalise_mlr_job(cohort: Cohort) -> PythonJob:
     )
     mlr_job.image(image=get_driver_image())
     return mlr_job
-
-
-def _run_command(
-    command: str | list[str],
-    capture_output: bool = False,
-    shell: bool = False,
-) -> subprocess.CompletedProcess:
-    """
-    Runs a subprocess command with robust error logging.
-    """
-    executable = '/bin/bash' if shell else None
-    cmd_str = command if isinstance(command, str) else ' '.join(command)
-
-    try:
-        logger.info(f'Running command: {cmd_str}')
-        return subprocess.run(
-            command,
-            check=True,
-            text=True,
-            capture_output=capture_output,
-            shell=shell,
-            executable=executable,
-        )
-    except subprocess.CalledProcessError as e:
-        logger.error(f'Command failed with return code {e.returncode}: {cmd_str}')
-        if e.stdout:
-            logger.error(f'STDOUT: {e.stdout.strip()}')
-        if e.stderr:
-            logger.error(f'STDERR: {e.stderr.strip()}')
-        # Re-raise as a generic exception to fail the job
-        raise ValueError('A subprocess command failed. See logs.') from e
-
-
-def _find_ica_file_path(parent_folder: str, file_name: str) -> str:
-    """
-    Finds a file in ICA and returns its full `details.path`.
-    """
-    command = [
-        'icav2',
-        'projectdata',
-        'list',
-        '--parent-folder',
-        parent_folder,
-        '--data-type',
-        'FILE',
-        '--file-name',
-        file_name,
-        '--match-mode',
-        'EXACT',
-        '-o',
-        'json',
-    ]
-    result = _run_command(command, capture_output=True)
-    try:
-        data = json.loads(result.stdout)
-        if not data.get('items'):
-            raise ValueError(f'No file found with name "{file_name}" in folder "{parent_folder}"')
-
-        file_path = data['items'][0].get('details', {}).get('path')
-        if not file_path:
-            raise ValueError(f'File "{file_name}" found, but it has no "details.path" in API response.')
-
-        logger.info(f'Found {file_name} at path: {file_path}')
-        return file_path
-
-    except json.JSONDecodeError:
-        logger.error(f'Failed to decode JSON from icav2 list command: {result.stdout}')
-        raise
-    except (ValueError, IndexError) as e:
-        logger.error(f'Error parsing icav2 list output for {file_name}: {e}')
-        raise
 
 
 def _submit_mlr_run(
@@ -123,20 +53,28 @@ def _submit_mlr_run(
     try:
         # --- 1. Authenticate ---
         # shell=True is required for the multi-line ICA_CLI_SETUP script
-        _run_command(ICA_CLI_SETUP, shell=True)
+        ica_utils.run_cli_command(ICA_CLI_SETUP, shell=True)
 
         # --- 2. Find input file paths ---
-        cram_path: str = _find_ica_file_path(ica_base_folder, f'{sg_name}.cram')
-        gvcf_path: str = _find_ica_file_path(ica_base_folder, f'{sg_name}.hard-filtered.gvcf.gz')
+        cram_path: str = ica_utils.find_ica_file_path_by_name(
+            ica_base_folder,
+            f'{sg_name}.cram',
+        )
+        gvcf_path: str = ica_utils.find_ica_file_path_by_name(
+            ica_base_folder,
+            f'{sg_name}.hard-filtered.gvcf.gz',
+        )
 
         cram_url: str = f'ica://OurDNA-DRAGEN-378/{cram_path.lstrip("/")}'
         gvcf_url: str = f'ica://OurDNA-DRAGEN-378/{gvcf_path.lstrip("/")}'
 
         # --- 3. Set ICA project context ---
-        _run_command(['icav2', 'projects', 'enter', mlr_project])
+        ica_utils.run_cli_command(
+            ['icav2', 'projects', 'enter', mlr_project],
+        )
 
         # --- 4. Download MLR config JSON ---
-        _run_command(
+        ica_utils.run_cli_command(
             [
                 'icav2',
                 'projectdata',
@@ -144,7 +82,7 @@ def _submit_mlr_run(
                 mlr_config_json,
                 local_config_path,
                 '--exclude-source-path',
-            ]
+            ],
         )
 
         # --- 5. Build and run the popgen-cli command ---
@@ -172,19 +110,26 @@ def _submit_mlr_run(
             '--analysis-instance-tier',
             config_retrieve(['ica', 'mlr', 'analysis_instance_tier']),
         ]
-        _run_command(submit_command)
+        ica_utils.run_cli_command(submit_command)  # <-- Changed
 
         # --- 6. Read the pipeline ID from the output JSON ---
-        output_json_path = os.path.join(sg_name, f'sample-{sg_name}-run-{sg_name}-mlr.json')
+        output_json_path = os.path.join(
+            sg_name,
+            f'sample-{sg_name}-run-{sg_name}-mlr.json',
+        )
         if not os.path.exists(output_json_path):
-            raise FileNotFoundError(f'popgen-cli did not produce expected output file: {output_json_path}')
+            raise FileNotFoundError(
+                f'popgen-cli did not produce expected output file: {output_json_path}',
+            )
 
         with open(output_json_path) as f:
             submission_data: dict[str, Any] = json.load(f)
 
         mlr_analysis_id = submission_data.get('id')
         if not mlr_analysis_id:
-            raise ValueError(f'Submission output file "{output_json_path}" is missing the "id" key.')
+            raise ValueError(
+                f'Submission output file "{output_json_path}" is missing the "id" key.',
+            )
 
         logger.info(f'MLR pipeline ID for {sg_name} is {mlr_analysis_id}')
         return mlr_analysis_id
@@ -219,7 +164,9 @@ def _run(
     """
     Calls the generic pipeline manager with settings for the MLR pipeline.
     """
-    ica_analysis_output_folder: str = config_retrieve(['ica', 'data_prep', 'output_folder'])
+    ica_analysis_output_folder: str = config_retrieve(
+        ['ica', 'data_prep', 'output_folder'],
+    )
     mlr_project: str = config_retrieve(['ica', 'projects', 'dragen_mlr'])
     dragen_align_project: str = config_retrieve(['ica', 'projects', 'dragen_align'])
     mlr_config_json: str = config_retrieve(['ica', 'mlr', 'config_json'])
@@ -229,7 +176,10 @@ def _run(
 
     def _create_submit_callable(sg_name: str) -> Callable[[], str]:
         """Creates a zero-argument callable for pipeline submission."""
-        output_prefix: str = f'ica://{dragen_align_project}/{BUCKET_NAME}/{config_retrieve(["ica", "data_prep", "output_folder"])}/{sg_name}'
+        output_prefix: str = (
+            f'ica://{dragen_align_project}/{BUCKET_NAME}/'
+            f'{config_retrieve(["ica", "data_prep", "output_folder"])}/{sg_name}'
+        )
 
         return partial(
             _submit_mlr_run,
