+++ src/dragen_align_pa/jobs/monitor_dragen_pipeline.py
@@ -5,7 +5,7 @@ from cpg_utils.config import config_retrieve
 from icasdk.apis.tags import project_analysis_api
 from loguru import logger
 
-from dragen_align_pa import utils
+from dragen_align_pa import ica_utils
 from dragen_align_pa.constants import ICA_REST_ENDPOINT
 
 
@@ -23,7 +23,7 @@ def run(ica_pipeline_id: str | dict[str, str], is_mlr: bool = False) -> str:
     Returns:
         str: The pipeline status
     """
-    secrets: dict[Literal['projectID', 'apiKey'], str] = utils.get_ica_secrets()
+    secrets: dict[Literal['projectID', 'apiKey'], str] = ica_utils.get_ica_secrets()
     project_id: str = secrets['projectID']
     api_key: str = secrets['apiKey']
 
@@ -31,15 +31,21 @@ def run(ica_pipeline_id: str | dict[str, str], is_mlr: bool = False) -> str:
     configuration.api_key['ApiKeyAuth'] = api_key
     pipeline_id: str = ica_pipeline_id['pipeline_id'] if isinstance(ica_pipeline_id, dict) else ica_pipeline_id
 
-    logger.info(f'Monitoring pipeline run {pipeline_id} which of type {type(pipeline_id)}')
+    logger.info(
+        f'Monitoring pipeline run {pipeline_id}',
+    )
     with icasdk.ApiClient(configuration=configuration) as api_client:
         api_instance = project_analysis_api.ProjectAnalysisApi(api_client)
         if not is_mlr:
             path_params: dict[str, str] = {'projectId': project_id}
         else:
-            path_params = {'projectId': config_retrieve(['ica', 'projects', 'dragen_mlr_project_id'])}
+            path_params = {
+                'projectId': config_retrieve(
+                    ['ica', 'projects', 'dragen_mlr_project_id'],
+                ),
+            }
 
-        return utils.check_ica_pipeline_status(
+        return ica_utils.check_ica_pipeline_status(
             api_instance=api_instance,
             path_params=path_params | {'analysisId': pipeline_id},
         )
